using int for timestamp

+++ OrderMatcher/OrderMatcher/CancelledOrder.cs
@@ -6,6 +6,6 @@
         public Quantity RemainingQuantity { get; set; }
         public Quantity RemainingOrderAmount { get; set; }
         public CancelReason CancelReason { get; set; }
-        public long Timestamp { get; set; }
+        public int Timestamp { get; set; }
     }
 }
