+++ OrderMatcher/OrderMatcher/Fill.cs
@@ -6,7 +6,7 @@
         public OrderId TakerOrderId { get; set; }
         public Price MatchRate { get; set; }
         public Quantity MatchQuantity { get; set; }
-        public long Timestamp { get; set; }
+        public int Timestamp { get; set; }
         public bool IncomingOrderFilled { get; set; }
     }
 }
